cloud/functions/create_avatar/requirements.txt
functions-framework==3.*
firebase-admin==6.*
google-genai==1.*
